requirements.txt
numpy
pandas
fastapi
uvicorn
python-dotenv
psycopg2
requests